HIVE-4041 [jira] Support multiple partitionings in a single Query (Harish Butani via Ashutosh Chauhan)
Summary:
add support for multiple partitioning

Currently we disallow queries if the partition specifications of all Wdw fns are not the same. We can relax this by generating multiple PTFOps based on the unique partitionings in a Query. For partitionings that only differ in sort, we can introduce a sort step in between PTFOps, which can happen in the same Reduce task.

Test Plan: added tests to windowing_multipartitioning.q

Reviewers: JIRA, ashutoshc

Reviewed By: ashutoshc

Differential Revision: https://reviews.facebook.net/D9381

git-svn-id: https://svn.apache.org/repos/asf/hive/branches/ptf-windowing@1458953 13f79535-47bb-0310-9956-ffa450edef68